Add image resizing to thresholding approach too Update layout text.
Note: this commit and the previous one also removed
the fetching of images via the fetch API as it was slow.
Instead now, when an the Image HTML element load its src
content, it gets encoded in base64 and sent to the
web Workers.
Also: it is important to base64-encode in JPEG format:
[using: canvas.toDataURL("image/jpeg", 1)], otherwise
statistical does not work properly.